% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{coef,fitsir-method}
\alias{coef,fitsir-method}
\title{Extract parameter of a fit}
\usage{
\S4method{coef}{fitsir}(object, type = c("raw", "trans", "summary"))
}
\arguments{
\item{object}{fitsir object}

\item{type}{type of parameter to be returned}
}
\description{
Extract parameter of a fit
}
\details{
\code{raw} returns unconstrained parameters; \code{trans} returns constrained parameters; and
\code{summary} returns summarized parameters.
}
\examples{
coef(ff)
coef(ff,"trans")
coef(ff,"summary")
}

