% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{coef.fitsir}
\alias{coef.fitsir}
\title{Coef method for fitsir objects}
\usage{
\S4method{coef}{fitsir}(object, type = c("raw", "trans", "summary"))
}
\arguments{
\item{object}{An object of class \code{fitsir}}

\item{type}{types of returned parameters}
}
\description{
Coef method for fitsir objects
}

