% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{dispersion,fitsir-method}
\alias{dispersion,fitsir-method}
\title{Find dispersion parameter}
\usage{
\S4method{dispersion}{fitsir}(object, dist = c("quasipoisson", "nbinom",
  "nbinom1"))
}
\arguments{
\item{object}{fitsir object}

\item{dist}{distribution}
}
\description{
Find dispersion parameter
}
\details{
\code{quasipoisson} returns the sum of pearson residuals divided by the degrees of freedom.
\code{nbinom} assumes quadratic mean-variance relation (var=mu+mu^2/k) and estimates k based on maximum likelihood.
\code{nbinom1} assumes linear mean-variance relation (var=(1+phi)mu) and estimates phi based on maximum likelihood.
}
