% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\name{dispersion}
\alias{dispersion}
\title{S4 method for finding a dispersion parameter}
\usage{
dispersion(object, ...)
}
\arguments{
\item{object}{an \code{R} object}

\item{...}{further arguments passed to methods}
}
\description{
S4 method for finding a dispersion parameter
}

