% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir.R
\name{dnbinom1}
\alias{dnbinom1}
\title{Reparameterized negative binomial with linear mean-variance relationship}
\usage{
dnbinom1(x, mu, tau, log = FALSE)
}
\arguments{
\item{x}{numeric value}

\item{tau}{numeric value}

\item{log}{(logical) return log-likelihood}

\item{mean}{mu of distribution}
}
\value{
likelihood or log-likelihood vector
}
\description{
Reparameterized negative binomial with linear mean-variance relationship
}
\details{
Variance is \code{tau * mu}
}

