% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{dnbinom2}
\alias{dnbinom2}
\title{Negative binomial distribution with profiled dispersion parameter}
\usage{
dnbinom2(x, mean, log = FALSE)
}
\arguments{
\item{x}{numeric value}

\item{mean}{mean of distribution}

\item{log}{(logical) return log-likelihood}
}
\value{
likelihood or log-likelihood vector
}
\description{
Negative binomial distribution with profiled dispersion parameter
}

