% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{dnorm2}
\alias{dnorm2}
\title{Normal distribution with profiled sd}
\usage{
dnorm2(x, mean, log = FALSE)
}
\arguments{
\item{x}{numeric value}

\item{mean}{mean of distribution}

\item{log}{(logical) return log-likelihood}
}
\value{
likelihood or log-likelihood vector
}
\description{
Normal distribution with profiled sd
}

