% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{Eval,loglik.fitsir-method}
\alias{Eval,loglik.fitsir-method}
\title{Evaluate a model}
\usage{
\S4method{Eval}{loglik.fitsir}(object, count, mean, par = NULL, ...)
}
\arguments{
\item{par}{a named vector (or list) of parameter values, if NULL, take from ...}

\item{...}{the input values and parameter values}

\item{data}{a dataframe object holding inut values, if NULL, take from ...}
}
\value{
numeric
}
\description{
Evaluate a model
}

