% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{findDeriv}
\alias{findDeriv}
\title{Derivative of negative log likelihood with respect to parameters}
\usage{
findDeriv(x, mean, nu, dist)
}
\description{
Derivative of negative log likelihood with respect to parameters
}

