% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{findHess}
\alias{findHess}
\title{find Hessian of SSQ wrt parameters}
\usage{
findHess(data, params)
}
\arguments{
\item{data}{data frame with tvec/count}

\item{params}{parameter vector}
}
\description{
find Hessian of SSQ wrt parameters
}

