% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{findSens}
\alias{findSens}
\title{returns}
\usage{
findSens(params, count, tvec = NULL, dist = c("norm", "pois"),
  type = c("prevalence", "incidence", "death"), debug = FALSE)
}
\description{
returns
}

