% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{findSSQ}
\alias{findSSQ}
\title{Calculate SSQ}
\usage{
findSSQ(data, params, incidence = FALSE, SSQonly = FALSE)
}
\arguments{
\item{data}{data for deterministic simulation (data frame with columns \code{tvec} and \code{count})}

\item{params}{parameter vector}

\item{incidence}{(logical) compute SSQ based on incidence?}

\item{SSQonly}{(logical) return only the sum of squares? (alternatively, return derivatives of SSQ with respect to parameters)}
}
\description{
Calculate SSQ
}

