% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-class.R
\docType{class}
\name{fitsir.mle2-class}
\alias{fitsir.mle2-class}
\title{Class "fitsir.mle2". 
Result of SIR model fitting based on Maximum Likelihood Estimation}
\description{
Class "fitsir.mle2". 
Result of SIR model fitting based on Maximum Likelihood Estimation
}
\seealso{
\code{\link{mle2-class}}
}

