% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{grad,loglik.fitsir-method}
\alias{grad,loglik.fitsir-method}
\title{the gradients w.r.t. to parameters
compute the gradient of the model as a function of the parameters}
\usage{
\S4method{grad}{loglik.fitsir}(object, count, mean, par, ...)
}
\arguments{
\item{par}{a named vector (or list) of parameters to compute the derivatives}

\item{...}{the input values and parameter values}

\item{data}{a dataframe object holding inut values, if NULL, take from ...}
}
\value{
a dataframe with each column as a partial derivative values
}
\description{
the gradients w.r.t. to parameters
compute the gradient of the model as a function of the parameters
}

