% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{grad.size}
\alias{grad.size}
\title{derivative of nbinom nll with respect to its dispersion parameter}
\usage{
grad.size(x, mean, size)
}
\description{
derivative of nbinom nll with respect to its dispersion parameter
}

