% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{graddsp}
\alias{graddsp}
\title{derivative of nbinom nll with respect to its dispersion parameter}
\usage{
graddsp(x, mean, log.dsp, dist)
}
\description{
derivative of nbinom nll with respect to its dispersion parameter
}

