% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{gradsize}
\alias{gradsize}
\title{derivative of nbinom nll with respect to its dispersion parameter}
\usage{
gradsize(x, mean, size)
}
\description{
derivative of nbinom nll with respect to its dispersion parameter
}

