% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{initialize,loglik.fitsir-method}
\alias{initialize,loglik.fitsir-method}
\title{the initializer for loglik.fitsir}
\usage{
\S4method{initialize}{loglik.fitsir}(.Object, name, model, count = "X", mean,
  par = NULL, keep_grad = TRUE, keep_hessian = FALSE)
}
\arguments{
\item{name}{name of the distribution}

\item{model}{the formula specifying the model}

\item{count}{observation variable name}

\item{mean}{mean variable name}

\item{par}{additional parameter names}

\item{keep_grad}{maintain the gradient as part of the model}

\item{keep_hessian}{maintain the hessian as part of the model}
}
\description{
the initializer for loglik.fitsir
}

