% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{initialize_hessian}
\alias{initialize_hessian}
\title{initialize the hessian model}
\usage{
initialize_hessian(dist = c("gaussian", "poisson", "quasipoisson", "nbinom",
  "nbinom1"))
}
\description{
initialize the hessian model
}
