% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{loglik.fitsir-class}
\alias{loglik.fitsir-class}
\title{Class representing log-likelihood models used to fit the SIR model}
\description{
Class representing log-likelihood models used to fit the SIR model
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the distribution}

\item{\code{expr}}{an expression specifying the model}

\item{\code{count}}{observation variable name}

\item{\code{mean}}{mean variable name}

\item{\code{par}}{additional parameter names}

\item{\code{grad}}{the gradient with respect to the parameters}

\item{\code{hessian}}{the Hessian of the model with respect}
}}

