% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{loglik_gaussian}
\alias{loglik_gaussian}
\title{gaussian log-likelihood model with sensitivity}
\format{An object of class \code{loglik.fitsir} of length 1.}
\usage{
loglik_gaussian
}
\description{
gaussian log-likelihood model with sensitivity
}
\keyword{datasets}
