% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{loglik_gaussian_base}
\alias{loglik_gaussian_base}
\title{gaussian log-likelihood base model}
\format{An object of class \code{loglik.fitsir} of length 1.}
\usage{
loglik_gaussian_base
}
\description{
gaussian log-likelihood base model
}
\keyword{datasets}

