% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{loglik_nbinom1}
\alias{loglik_nbinom1}
\title{negative binomial 1 log-liklihood model with sensitivity}
\format{An object of class \code{loglik.fitsir} of length 1.}
\usage{
loglik_nbinom1
}
\description{
negative binomial 1 log-liklihood model with sensitivity
}
\keyword{datasets}

