% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{loglik_nbinom1_base}
\alias{loglik_nbinom1_base}
\title{negative binomial 1 log-liklihood base model}
\format{An object of class \code{loglik.fitsir} of length 1.}
\usage{
loglik_nbinom1_base
}
\description{
negative binomial 1 log-liklihood base model
}
\keyword{datasets}

