% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{loglik_nbinom_base}
\alias{loglik_nbinom_base}
\title{negative binomial log-likelihood base model}
\format{An object of class \code{loglik.fitsir} of length 1.}
\usage{
loglik_nbinom_base
}
\description{
negative binomial log-likelihood base model
}
\keyword{datasets}
