% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{loglik_poisson_base}
\alias{loglik_poisson_base}
\title{poisson log-likelihood base model}
\format{An object of class \code{loglik.fitsir} of length 1.}
\usage{
loglik_poisson_base
}
\description{
poisson log-likelihood base model
}
\keyword{datasets}
