% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{mle.size}
\alias{mle.size}
\title{Maximum likelihood estimate of negative binomial dispersion parameter}
\usage{
mle.size(x, mean)
}
\description{
Maximum likelihood estimate of negative binomial dispersion parameter
}

