% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{mlesize}
\alias{mlesize}
\title{Maximum likelihood estimate of negative binomial dispersion parameter}
\usage{
mlesize(x, mean)
}
\description{
Maximum likelihood estimate of negative binomial dispersion parameter
}

