% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{phila1918}
\alias{phila1918}
\title{1918 Philadelphia flu data set}
\format{A data frame with two elements}
\usage{
data(phila1918)
}
\description{
1918 Philadelphia flu data set
}
\keyword{datasets}
