% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{plot,fitsir,missing-method}
\alias{plot,fitsir,missing-method}
\alias{plot,fitsir-method}
\title{Plot a fitsir object}
\usage{
\S4method{plot}{fitsir,missing}(x, level, method = c("delta", "mvrnorm",
  "wmvrnorm"), main, xlim, ylim, xlab, ylab, add = FALSE,
  col.traj = "black", lty.traj = 1, col.conf = "red", lty.conf = 4, ...)
}
\arguments{
\item{x}{fitsir object}

\item{level}{the confidence level required}

\item{method}{confidence interval method}

\item{main}{an overall title for the plot}

\item{xlim}{the x limit of the plot}

\item{ylim}{the y limit of the plot}

\item{xlab}{a label for the x axis}

\item{ylab}{a label for the y axis}

\item{add}{(logical) add to an existing plot?}

\item{col.traj}{colour of the estimated trajectory}

\item{lty.traj}{line type of the estimated trajectory}

\item{col.conf}{colour of the confidence intervals}

\item{lty.conf}{line type of the confidence intervals}

\item{...}{additional arguments to be passed on to the plot function}
}
\description{
Plot a fitsir object
}
\examples{
harbin2 <- setNames(harbin, c("times", "count"))
ss <- startfun(harbin2, type="death")
ff <- fitsir(harbin2, start=ss, type="death", method="BFGS")
plot(ff)

ff2 <- fitsir(harbin2, start=c(ss, ll.k=5), type="death", dist="nbinom", method="BFGS")
ff3 <- fitsir(harbin2, start=ss, type="death", dist="quasipoisson", method="BFGS")
plot(ff2, level=0.95, col.traj="red", main="Negative binomial error vs. Quasipoisson error CIs")
plot(ff3, add=TRUE, level=0.95, col.traj="blue", col.conf="blue")
legend(2, 270, legend = c("NB2", "Quasipoisson"), col=c("red", "blue"), lty=1)
}

