% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{plot.fitsir}
\alias{plot.fitsir}
\title{S4 plot method for fitsir objects}
\usage{
\S4method{plot}{fitsir,missing}(x, level, main, xlim, ylim, xlab, ylab,
  add = FALSE, col.traj = "black", lty.traj = 1, col.conf = "red",
  lty.conf = 4, ...)
}
\arguments{
\item{x}{a fitsir object}

\item{level}{the confidence level required}

\item{main}{main title}

\item{xlab}{x label}

\item{ylab}{y label}

\item{add}{(logical) add trajectories to an existing figure}

\item{col.traj}{color for main trajectory}

\item{lty.traj}{line type for main trajectory}

\item{col.conf}{color for trajectories based on confidence intervals}

\item{lty.conf}{line type for trajectories based on confidence intervals}

\item{...}{additional arguments}
}
\description{
S4 plot method for fitsir objects
}
\examples{
harbin2 <- setNames(harbin, c("times", "count"))
ff <- fitsir(harbin2, type="death", method="BFGS")
plot(ff)

ff2 <- fitsir(harbin2, type="death", dist="nbinom", method="BFGS")
ff3 <- fitsir(harbin2, type="death", dist="quasipoisson", method="BFGS")
plot(ff2, level=0.95, col.traj="red", main="Negative binomial error vs. Quasipoisson error CIs")
plot(ff3, add=TRUE, level=0.95, col.traj="blue", col.conf="blue")
legend(2, 270, legend = c("NB2", "Quasipoisson"), col=c("red", "blue"), lty=1)
}

