% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{predict,fitsir-method}
\alias{predict,fitsir-method}
\title{Forecast from an SIR fit and find confidence interval}
\usage{
\S4method{predict}{fitsir}(object, level, times, method = c("delta",
  "mvrnorm", "wmvrnorm"), debug = FALSE)
}
\arguments{
\item{object}{fitsir object}

\item{level}{the confidence level required}

\item{times}{time vector to predict over. Default is set to the time frame of the data.}

\item{method}{confidence interval method. Default is set to Delta method.}

\item{debug}{print debugging output?}
}
\description{
Forecast from an SIR fit and find confidence interval
}
\details{
See vignette for different methods: \code{vignette("details", package="fitsir")}
}
\examples{
predict(ff, level=0.95)
}

