% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{predict.fitsir}
\alias{predict.fitsir}
\title{S4 predict method for fitsir objects}
\usage{
\S4method{predict}{fitsir}(object, level, times)
}
\arguments{
\item{object}{a fitsir object}

\item{level}{the confidence level required}

\item{times}{new time vector}
}
\description{
S4 predict method for fitsir objects
}
\examples{
ff <- fitsir(harbin, type="death", method="BFGS", tcol="week", icol="Deaths")
predict(ff, level=0.95)
}

