% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{residuals,fitsir-method}
\alias{residuals,fitsir-method}
\title{Find residuals between the fit and the data}
\usage{
\S4method{residuals}{fitsir}(object, type = c("pearson", "raw"))
}
\arguments{
\item{object}{fitsir object}

\item{type}{type of residuals. Default is set to pearson residuals.}
}
\description{
Find residuals between the fit and the data
}
\details{
\code{pearson} returns \eqn{(X_i - \mu_i)^2/\mu_i} and \code{raw} retrns \eqn{X_i -\mu_i}, 
where \eqn{X_i} is the observed counts and \eqn{\mu_i} is the expected coutns.
}
\examples{
harbin2 <- setNames(harbin, c("times", "count"))
ss <- startfun(harbin2, type="death")
ff <- fitsir(harbin2, start=ss, type="death", method="BFGS")
residuals(ff)
}

