% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{residuals.fitsir}
\alias{residuals.fitsir}
\title{S4 residuals method for fitsir objects}
\usage{
\S4method{residuals}{fitsir}(object, type = c("pearson", "raw"))
}
\arguments{
\item{object}{a fitsir object}

\item{type}{type of residuals}
}
\description{
S4 residuals method for fitsir objects
}
\examples{
ff <- fitsir(harbin, type="death", method="BFGS", tcol="week", icol="Deaths")
residuals(ff)
}

