% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{show,summary.fitsir-method}
\alias{show,summary.fitsir-method}
\title{Show summary of a fit}
\usage{
\S4method{show}{summary.fitsir}(object)
}
\arguments{
\item{summary.fitsir}{object}
}
\description{
Show summary of a fit
}

