% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{SIR.detsim.hessian}
\alias{SIR.detsim.hessian}
\title{integrate sensitivities}
\usage{
SIR.detsim.hessian(t, params)
}
\arguments{
\item{t}{time vector}

\item{params}{parameter vector}
}
\description{
integrate sensitivities
}

