% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SIR.detsim}
\alias{SIR.detsim}
\title{deterministic trajectory of SIR}
\usage{
SIR.detsim(t, params, func = SIR.grad)
}
\arguments{
\item{t}{time vector}

\item{params}{parameter vector (beta, gamma, N, i0)}

\item{func}{gradient function}
}
\description{
deterministic trajectory of SIR
}

