% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad.R
\name{SIR.grad}
\alias{SIR.grad}
\alias{SIR.grad.sens}
\title{gradient function (solves with {S,log(I)} for stability)}
\usage{
SIR.grad(t, x, params)

SIR.grad.sens(t, x, params)
}
\arguments{
\item{t}{time vector}

\item{x}{state vector (with names \code{S}, \code{logI})}

\item{params}{parameter vector (with names \code{beta} (scaled transmission rate), \code{N} (population size), \code{gamma} (recovery/removal rate)}
}
\value{
gradient vector for a simple SIR model
}
\description{
gradient function (solves with {S,log(I)} for stability)
}
