% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{SIR.hessian}
\alias{SIR.hessian}
\title{find Hessian}
\usage{
SIR.hessian(data, params, dist = c("gaussian", "poisson", "quasipoisson",
  "nbinom", "nbinom1"), type = c("prevalence", "incidence", "death"),
  tcol = "times", icol = "count")
}
\arguments{
\item{data}{data frame with tvec/count}

\item{params}{parameter vector}

\item{dist}{conditional distribution of reported data}

\item{type}{type of reported data}

\item{tcol}{column name for time variable}

\item{icol}{column name for count variable}
}
\description{
find Hessian
}
