% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startfun.R
\name{smooth.spline2}
\alias{smooth.spline2}
\title{Fit a spline to an epidemic data}
\usage{
smooth.spline2(times, count, itmax = 100, relpeakcrit = 0.1)
}
\arguments{
\item{times}{time vector}

\item{count}{data (epidemic counts for each time period)}

\item{itmax}{maximum number of iterations}

\item{relpeakcrit}{critical relative peak value to test for single peak}
}
\description{
Fit a spline to an epidemic data
}

