% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir.R
\name{summarize.pars}
\alias{summarize.pars}
\alias{summarize.pars.jacobian}
\title{Summarize parameters}
\usage{
\method{summarize}{pars}(params)

\method{summarize}{pars.jacobian}(params)
}
\arguments{
\item{params}{parameter vector (log.beta, log.gamma, logit.i)}
}
\description{
Generate meaningful epidemiological summary statistics (R0, r, infectious period, I(0)) from a set of epidemic parameters
}
