% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{summary,fitsir-method}
\alias{summary,fitsir-method}
\title{Summarize the fit}
\usage{
\S4method{summary}{fitsir}(object)
}
\arguments{
\item{object}{fitsir object}
}
\description{
Summarize the fit
}
