% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-class.R
\docType{class}
\name{summary.fitsir-class}
\alias{summary.fitsir-class}
\title{Class "summary.fitsir".
Summary of SIR model fit}
\description{
Class "summary.fitsir".
Summary of SIR model fit
}
\section{Slots}{

\describe{
\item{\code{call}}{Objecto of class "\code{language}".
The call that generated the "fitsir" object.}

\item{\code{coef}}{Object of class "\code{matrix}".
Estimated coefficients and standard errors.}

\item{\code{summary}}{Object of class "\code{matrix}".
Summary of estimated coefficients and standard errors.}

\item{\code{m2logL}}{Object of class "\code{numeric}".
Minus twice the log likelihood.}
}}
\seealso{
\code{\link{summary.mle2-class}}
}

