% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir.R
\name{trans.pars}
\alias{trans.pars}
\title{transform parameters}
\usage{
trans.pars(params, scale = c("constrained", "unconstrained"))
}
\arguments{
\item{params}{parameter vector containing \code{log.beta}, \code{log.gamma}, \code{log.N}, \code{logit.i}}
}
\value{
params transformed parameter vector containing \code{beta}, \code{gamma}, \code{N}, \code{i0}
}
\description{
transform parameters
}
