% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{Transform,loglik.fitsir-method}
\alias{Transform,loglik.fitsir-method}
\title{Transform the model}
\usage{
\S4method{Transform}{loglik.fitsir}(object, transforms = NULL, name,
  count = "X", mean, par, keep_grad = TRUE, keep_hessian = FALSE)
}
\arguments{
\item{object}{object}

\item{transforms}{list of formulas specifying transformations}

\item{count}{observation variable name}

\item{mean}{mean variable name}

\item{par}{additional parameter names}

\item{keep_grad}{maintain the gradient as part of the model}

\item{keep_hessian}{maintain the hessian as part of the model}
}
\value{
loglik.fitsir object
}
\description{
Transform the model
}

