% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alrinv}
\alias{alr}
\alias{alrinv}
\title{additive log-ratio transformation and inverse}
\usage{
alrinv(x)

alr(x)
}
\arguments{
\item{x}{value to transform (or inverse-transform)}
}
\description{
additive log-ratio transformation and inverse
}

