% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{derivfun}
\alias{derivfun}
\title{Derivative of negative log likelihood with respect to parameters}
\usage{
derivfun(x, mean, log.dsp = NULL, nu, dist)
}
\description{
Derivative of negative log likelihood with respect to parameters
}

