% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dnorm2}
\alias{dnorm2}
\title{Normal distribution with profiled sd}
\usage{
dnorm2(x, mean, log = FALSE)
}
\arguments{
\item{x}{numeric value}

\item{mean}{mean of distribution}

\item{log}{(logical) return log-likelihood?}
}
\value{
likelihood or log-likelihood vector
}
\description{
Normal distribution with profiled sd
}

