% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{Eval,loglik.fitsir-method}
\alias{Eval,loglik.fitsir-method}
\title{Evaluate log likelihood model}
\usage{
\S4method{Eval}{loglik.fitsir}(object, count, mean, par = NULL, ...)
}
\arguments{
\item{object}{object to be evaluated}

\item{count}{observations}

\item{mean}{mean values}

\item{par}{additional parameters}

\item{...}{other values if required}
}
\value{
numeric
}
\description{
Evaluate log likelihood model
}
