% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{findHess}
\alias{findHess}
\title{find Hessian}
\usage{
findHess(data, params, dist = c("gaussian", "poisson", "quasipoisson",
  "nbinom", "nbinom1"))
}
\arguments{
\item{data}{data frame with tvec/count}

\item{params}{parameter vector}
}
\description{
find Hessian
}

