% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-class.R, R/fitsir-methods.R
\docType{class}
\name{fitsir-class}
\alias{coef,fitsir-class}
\alias{coef,fitsir-method}
\alias{dispersion,fitsir-class}
\alias{dispersion,fitsir-method}
\alias{fitsir-class}
\alias{plot,fitsir,missing-method}
\alias{plot,fitsir-class}
\alias{predict,fitsir-class}
\alias{predict,fitsir-method}
\alias{residuals,fitsir-class}
\alias{residuals,fitsir-method}
\alias{summary,fitsir-class}
\alias{summary,fitsir-method}
\title{Class "fitsir". 
Result of SIR model fitting based on Maximum Likelihood Estimation}
\usage{
\S4method{plot}{fitsir,missing}(x, level, method = c("delta", "mvrnorm",
  "wmvrnorm", "sobol"), main, xlim, ylim, xlab, ylab, add = FALSE,
  col.traj = "black", lty.traj = 1, col.conf = "red", lty.conf = 4, ...)

\S4method{coef}{fitsir}(object, type = c("raw", "trans", "summary"))

\S4method{predict}{fitsir}(object, level, times, method = c("delta",
  "mvrnorm", "wmvrnorm", "sobol"), debug = FALSE)

\S4method{residuals}{fitsir}(object, type = c("pearson", "raw"))

\S4method{dispersion}{fitsir}(object, dist = c("quasipoisson", "nbinom",
  "nbinom1"))

\S4method{summary}{fitsir}(object, ...)
}
\description{
Class "fitsir". 
Result of SIR model fitting based on Maximum Likelihood Estimation
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot deterministic trajectory

\item \code{coef}: extract coefficients

\item \code{predict}: predict deterministic trajectory

\item \code{residuals}: calculate residuals between data and deterministic trajectory

\item \code{dispersion}: calculate dispersion parameter

\item \code{summary}: summarize fit using meaningful parameters
}}
\examples{
harbin2 <- setNames(harbin, c("times", "count"))
ff <- fitsir(harbin2, type="death", method="BFGS")
plot(ff)

ff2 <- fitsir(harbin2, type="death", dist="nbinom", method="BFGS")
ff3 <- fitsir(harbin2, type="death", dist="quasipoisson", method="BFGS")
plot(ff2, level=0.95, col.traj="red", main="Negative binomial error vs. Quasipoisson error CIs")
plot(ff3, add=TRUE, level=0.95, col.traj="blue", col.conf="blue")
legend(2, 270, legend = c("NB2", "Quasipoisson"), col=c("red", "blue"), lty=1)

coef(ff)
coef(ff,"trans")
coef(ff,"summary")

predict(ff, level=0.95)


residuals(ff)

}
\seealso{
\code{\link{mle2-class}}
}

