% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitsir}
\alias{fitsir}
\title{fitting function}
\usage{
fitsir(data, method = "Nelder-Mead", control = list(maxit = 1e+05),
  start = startfun(), debug = FALSE)
}
\arguments{
\item{data}{data frame with columns \code{tvec} and \code{count}}

\item{method}{optimization method}

\item{control}{control parameters for optimization}

\item{start}{starting parameters}

\item{debug}{print debugging output?}
}
\description{
fitting function
}

