% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{fitsir}
\alias{fitsir}
\title{fitting function}
\usage{
fitsir(data, start = startfun(), dist = c("norm", "pois", "qpois",
  "nbinom"), type = c("prevalence", "incidence", "death"),
  method = c("Nelder-Mead", "BFGS", "SANN"), control = list(maxit = 1e+05),
  timescale = NULL, verbose = FALSE, debug = FALSE)
}
\arguments{
\item{data}{data frame with columns \code{tvec} and \code{count}}

\item{start}{starting parameters}

\item{dist}{conditional distribution of reported data}

\item{type}{type of reported data}

\item{method}{optimization method}

\item{control}{control parameters for optimization}

\item{debug}{print debugging output?}
}
\description{
fitting function
}
\examples{
bombay2 <- setNames(bombay,c("tvec","count"))
(f1 <- fitsir(bombay2, type="death"))
plot(f1)
ss <- SIR.detsim(bombay2$tvec,trans.pars(coef(f1)))
cc <- bombay2$count

## CRUDE R^2 analogue (don't trust it too far! only works if obs always>0)
mean((1-ss/cc)^2)
}

