% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{grad,loglik.fitsir-method}
\alias{grad,loglik.fitsir-method}
\title{Evaluate the gradient of a model}
\usage{
\S4method{grad}{loglik.fitsir}(object, count, mean, par, var, ...)
}
\arguments{
\item{object}{object to be evaluated}

\item{count}{observations}

\item{mean}{mean values}

\item{par}{additional parameters}

\item{var}{character vector specifying names of partial derivative parameters}

\item{...}{other values if required}
}
\value{
a list with each element as a partial derivative values
}
\description{
Evaluate the gradient of a model
}

