% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{harbin}
\alias{harbin}
\title{1912 Harbin plague data set}
\format{A data frame with two elements}
\usage{
data(harbin)
}
\description{
1912 Harbin plague data set
}
\references{
Dietz, K. 2009. "Epidemics: the fitting of the first dynamic models to data". Journal of Contemporary Mathematical Analysis (Armenian Academy of Sciences), 44(2), 97-104.
}
\keyword{datasets}

