% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{hessian,loglik.fitsir-method}
\alias{hessian,loglik.fitsir-method}
\title{the hessian w.r.t. to parameters
compute the hessian of the model as a function of the parameters}
\usage{
\S4method{hessian}{loglik.fitsir}(object, data = NULL, par, var, ...)
}
\arguments{
\item{data}{a dataframe object holding inut values, if NULL, take from ...}

\item{par}{a named vector (or list) of parameters to compute the hessians}

\item{...}{the input values and parameter values}
}
\value{
a dataframe with each column as a partial derivative values
}
\description{
the hessian w.r.t. to parameters
compute the hessian of the model as a function of the parameters
}

