% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir.R
\name{minusloglfun}
\alias{minusloglfun}
\title{Negative log likelihood functions}
\usage{
minusloglfun(x, mean, par, dist)
}
\arguments{
\item{x}{vector of observations}

\item{mean}{vector of means}

\item{par}{additional parameter}

\item{dist}{conditional distribution of reported data}
}
\description{
Negative log likelihood functions
}

