% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir.R
\name{mledsp}
\alias{mledsp}
\title{Maximum likelihood estimate of negative binomial dispersion parameter}
\usage{
mledsp(x, mean, dist = c("nbinom", "nbinom1"))
}
\arguments{
\item{x}{vector of counts}

\item{mean}{mean of the distribution}

\item{dist}{conditional distribution of reported data}
}
\description{
Maximum likelihood estimate of negative binomial dispersion parameter
}
