% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{mledsp}
\alias{mledsp}
\title{Maximum likelihood estimate of negative binomial dispersion parameter}
\usage{
mledsp(x, mean, dist)
}
\description{
Maximum likelihood estimate of negative binomial dispersion parameter
}

