% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir-methods.R
\docType{methods}
\name{plot.fitsir}
\alias{plot.fitsir}
\title{Plotting method for fitsir objects}
\usage{
\S4method{plot}{fitsir,missing}(x, level, main, xlim, ylim, xlab, ylab,
  add = FALSE, col.traj = "black", lty.traj = 1, col.conf = "red",
  lty.conf = 4, ...)
}
\arguments{
\item{x}{An object of class \code{fitsir}}

\item{level}{the confidence level required}

\item{main}{main title}

\item{xlab}{x label}

\item{ylab}{y label}

\item{add}{(logical) add trajectories to an existing figure}

\item{col.traj}{color for main trajectory}

\item{lty.traj}{line type for main trajectory}

\item{col.conf}{color for trajectories based on confidence intervals}

\item{lty.conf}{line type for trajectories based on confidence intervals}

\item{...}{additional arguments}
}
\description{
Plotting method for fitsir objects
}
\examples{
harbin2 <- setNames(harbin, c("tvec", "count"))
ff <- fitsir(harbin2, type="death")
plot(ff)

plot(ff, level=0.95, log="y")
ff2 <- fitsir(harbin2, type="death", method="BFGS")
plot(ff2, add=TRUE, col.traj="blue")

plot(ff2, level=0.8, col.traj="red", log="y")
ff3 <- fitsir(harbin2, type="death", dist="pois", method="BFGS")
plot(ff3, add=TRUE, level=0.8, col.traj="blue", col.conf="blue")

ff4 <- fitsir(harbin2, type="death", dist="nbinom")
plot(ff4, level=0.8, log="y")
}

