% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{SIR.detsim}
\alias{SIR.detsim}
\title{deterministic trajectory of SIR}
\usage{
SIR.detsim(t, params, func = SIR.grad)
}
\arguments{
\item{t}{time vector}

\item{params}{parameter vector (beta, gamma, N, i0)}

\item{func}{gradient function}
}
\description{
deterministic trajectory of SIR
}
\examples{
pars <- c(beta=0.2,gamma=0.1,N=1000,i0=0.01)
tvec <- seq(0,200,by=0.01)
ss <- SIR.detsim(tvec,pars)
plot(tvec,ss,type="l",xlab="time",ylab="infected")
}

