% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir.R
\name{SIR.detsim}
\alias{SIR.detsim}
\title{deterministic trajectory of SIR}
\usage{
SIR.detsim(t, params, type = c("prevalence", "incidence", "death"),
  grad = FALSE)
}
\arguments{
\item{t}{time vector}

\item{params}{parameter vector (beta, gamma, N, i0)}

\item{type}{type of count data}

\item{grad}{(logical) return gradient with respect to unconstrained parameters}
}
\description{
deterministic trajectory of SIR
}
\examples{
pars <- c(beta=0.4,gamma=0.2,N=5000,i0=0.001)
times <- 0:50
ss.p <- SIR.detsim(times,pars)
ss.i <- SIR.detsim(times,pars,type="incidence")
ss.d <- SIR.detsim(times,pars,type="death")
matplot(data.frame(ss.p,ss.i,ss.d),type = "l",xlab="time",ylab="count")
legend(x=0,y=800,col=1:3,lty=1:3,legend=c("prevalence","incidence","death"))
all.equal(cumsum(c(5, ss.i[-length(ss.i)])) - cumsum(c(0, ss.d[-length(ss.d)])), ss.p)
}
