% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{SIR.hessian}
\alias{SIR.hessian}
\title{find Hessian}
\usage{
SIR.hessian(data, params, dist = c("gaussian", "poisson", "quasipoisson",
  "nbinom", "nbinom1"), tcol = "times", icol = "count")
}
\arguments{
\item{data}{data frame with tvec/count}

\item{params}{parameter vector}
}
\description{
find Hessian
}
