% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SIR.logLik}
\alias{SIR.logLik}
\title{Normal log-likelihood for SIR trajectory}
\usage{
SIR.logLik(params, count, tvec = NULL, debug = FALSE)
}
\arguments{
\item{params}{}

\item{count}{data (epidemic counts for each time period)}

\item{tvec}{time vector}

\item{debug}{print debugging output?}
}
\description{
Normal log-likelihood for SIR trajectory
}

