% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{SIR.logLik}
\alias{SIR.logLik}
\title{Log-likelihood for SIR trajectory}
\usage{
SIR.logLik(params, count, tvec = NULL, dist = c("norm", "pois", "nbinom"),
  type = c("prevalence", "incidence", "death"), debug = FALSE)
}
\arguments{
\item{params}{parameter vector (log.N0, logit.i0, log.beta, log.gamma)}

\item{count}{data (epidemic counts for each time period)}

\item{tvec}{time vector}

\item{dist}{conditional distribution of reported data}

\item{type}{type of reported data}

\item{debug}{print debugging output?}
}
\description{
Log-likelihood for SIR trajectory
}

