% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsir.R
\name{SIR.sensitivity}
\alias{SIR.sensitivity}
\title{Gradient of negative log likelihood with respect to each parameters}
\usage{
SIR.sensitivity(params, count, times = NULL, dist = c("gaussian", "poisson",
  "quasipoisson", "nbinom", "nbinom1"), type = c("prevalence", "incidence",
  "death"), debug = FALSE)
}
\arguments{
\item{params}{parameter vector (log.N, logit.i, log.beta, log.gamma)}

\item{count}{data (epidemic counts for each time period)}

\item{times}{time vector}

\item{dist}{conditional distribution of reported data}

\item{type}{type of reported data}

\item{debug}{print debugging output?}
}
\description{
Gradient of negative log likelihood with respect to each parameters
}
\examples{
fitsir:::SIR.sensitivity(c(log.beta=2,log.gamma=0,logit.i=-3,log.N=4),
                 count=c(1,2,4,7,3),
                 times=1:5,
                 dist="gaussian",
                 type="prevalence")

}

