% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{SIR.sensitivity}
\alias{SIR.sensitivity}
\title{Gradient of negative log likelihood with respect to each parameters}
\usage{
SIR.sensitivity(params, count, tvec = NULL, dist = c("norm", "pois",
  "qpois", "nbinom"), type = c("prevalence", "incidence", "death"),
  debug = FALSE)
}
\arguments{
\item{params}{parameter vector (log.N0, logit.i0, log.beta, log.gamma)}

\item{count}{data (epidemic counts for each time period)}

\item{tvec}{time vector}

\item{dist}{conditional distribution of reported data}

\item{type}{type of reported data}
}
\description{
Gradient of negative log likelihood with respect to each parameters
}

