% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startfun.R
\name{startfun}
\alias{startfun}
\title{Starting function}
\usage{
startfun(data, type = c("prevalence", "incidence", "death"), tcol = "times",
  icol = "count", itmax = 100, relpeakcrit = 0.1)
}
\arguments{
\item{data}{data frame with columns \code{times} and \code{count}}

\item{type}{epidemic data type}

\item{tcol}{column name for time variable}

\item{icol}{column name for count variable}

\item{itmax}{maximum number of iterations in \code{\link{smooth.spline2}}}

\item{relpeakcrit}{critical relative peak value used in \code{\link{smooth.spline2}}}
}
\description{
Starting function
}

