% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startfun.R
\name{startfun}
\alias{startfun}
\title{Starting function}
\usage{
startfun(data = NULL, type = c("prevalence", "incidence", "death"),
  log.beta = log(0.12), log.gamma = log(0.09), log.N = log(10000),
  logit.i = qlogis(0.01), itmax = 100, relpeakcrit = 0.1)
}
\arguments{
\item{data}{data frame with columns \code{tvec} and \code{count}}

\item{type}{epidemic data type}

\item{log.beta}{log of per capita transmission rate}

\item{log.gamma}{log of recovery/removal rate}

\item{log.N}{log of population size}

\item{logit.i}{logit of initial proportion infectious}

\item{itmax}{maximum number of iterations in \code{\link{spline2}}}

\item{relpeakcrit}{critical relative peak value used in \code{\link{spline2}}}
}
\description{
Starting function
}

