% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{summarize.pars}
\alias{summarize.pars}
\title{Summarize parameters}
\usage{
summarize.pars(params)
}
\arguments{
\item{params}{parameter vector (log.beta, log.gamma, logit.i)}
}
\description{
Generate meaningful epidemiological summary statistics (R0, r, infectious period, I(0) from a set of epidemic parameters
}

