% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitSIR_funs.R
\name{startfun}
\alias{startfun}
\alias{trans.pars}
\title{Starting function}
\usage{
startfun(log.beta = log(0.12), log.gamma = log(0.09), log.N = log(10000),
  logit.i = qlogis(0.01))

trans.pars(params)
}
\arguments{
\item{log.beta}{log of per capita transmission rate}

\item{log.gamma}{log of recovery/removal rate}

\item{log.N}{log of population size}

\item{logit.i}{logit of initial proportion infectious}

\item{params}{parameter vector containing \code{log.beta}, \code{log.gamma}, \code{log.N}, \code{logit.i}}
}
\value{
params transformed parameter vector containing \code{beta}, \code{gamma}, \code{N}, \code{s0}, \code{i0}
}
\description{
Starting function

transform parameters log->exp or alr->raw
*assume* R=0/S=N at start of epidemic
}

